+++ CHANGELOG.md
@@ -7,5 +7,12 @@ The changelog is available [on GitHub][2].
 
 * Initially created.
 
+## 0.0.1.0
+
+* Dropped (minimum required) cabal version to improve compatibility
+* Added functions to generate A) invertible random field element, and B) random Pallas/Vesta curve point
+* Added `Bounded` instances to field elements
+
+
 [1]: https://pvp.haskell.org
 [2]: https://github.com/nccgroup/pasta-curves/releases